Added a canvas resizing functionality
Struggled with this part the most as I wasnt realizing that after I
remove the already present canvas that I need to reselect all the canvas
divs again with another querySelectorAll statement so that the evenListe
ners could actually see the new divs that were created.

Users can now select the Canvas Size button to input a new option that
will then remove the old canvas, create a new one, querySelectorAll the
new divs of the new canvas, and applies new evenListeners on each one
for the draw modes.