Finished styling the page overall.
Also added JS functionality that adds active response when a draw mode is selected inverting the color of the text and button background to show that is the active button currently, does not effect the clear, gridlines, and canvas button. Also removed the 'canvasSize' div as I ended up not really needed it. In the future if I wasnt to add a canvas size slider I'd just put it below the actual canvas I think. Also was able to add the shading draw option that progressively shades a div by 10% for each interaction with that div (just white to black). This required having the JS actively set the background color of the divs to white each times a canvas is made since thats what the conditional initially looks for at step one of the shading.